Ignore unavailable and unknown states (#31)
Some device trackers (especially the newer entity-based variety, like Life360 just became in 2022.7) and binary sensors can temporarily have unavailable or unknown states. If this is true on two consecutive updates for an entity the composite tracker will stop listening to it.

This change causes the composite tracker to ignore unavailable and unknown states of the input entities, similar to what the Person integration does.